requirements.txt
absl-py==2.3.1
albucore==0.0.24
albumentations==2.0.8
altair==5.5.0
annotated-types==0.7.0
astunparse==1.6.3
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.8
coloredlogs==15.0.1
contourpy==1.3.0
cycler==0.12.1
Cython==3.1.2
deepface==0.0.93
dlib==20.0.0
easydict==1.13
eval_type_backport==0.2.2
face-recognition==1.3.0
face_recognition_models==0.3.0
filelock==3.18.0
fire==0.7.0
Flask==3.1.1
flask-cors==6.0.1
flatbuffers==25.2.10
fonttools==4.59.0
fsspec==2025.7.0
gast==0.6.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-pasta==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.73.1
gunicorn==23.0.0
h5py==3.14.0
httplib2==0.22.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
insightface==0.7.3
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
keras==3.10.0
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
mtcnn==0.1.1
namex==0.1.0
narwhals==1.47.1
networkx==3.2.1
numpy==2.0.2
oauth2client==4.1.3
onnx==1.18.0
onnxruntime==1.19.2
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
prettytable==3.16.0
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
PyDrive==1.3.1
Pygments==2.19.2
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
retina-face==0.0.17
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
simsimd==6.5.0
six==1.17.0
smmap==5.0.2
soupsieve==2.7
streamlit==1.47.0
stringzilla==3.12.5
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2024.8.30
toml==0.10.2
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.23.0
